Move statistic chart to a stimulus chart controller
