+++ assets/app.js
@@ -1,1 +1,2 @@
-import './scripts/statistics_chart.js';
+// App entrypoint - Leave empty unless need to override entrypoint from BootstrapAdminUi
+
