+++ templates/dashboard/index/content/statistics_chart.html.twig
@@ -58,6 +58,13 @@
                 </button>
             </div>
         </div>
-        <div id="statistics-chart" class="chart-lg" data-intervals="{{ statistics_data.talks_summary.intervals|json_encode }}" data-talks="{{ statistics_data.talks_summary.talks|json_encode }}"></div>
+        <div
+            id="statistics-chart"
+            class="chart-lg"
+            {{ stimulus_controller('chart', {
+                intervals: statistics_data.talks_summary.intervals,
+                talks: statistics_data.talks_summary.talks,
+            }) }}
+        ></div>
     </div>
 </div>
